Refactor components folder structure
