+++ src/components/modal/modal-ingredient-details/modal-ingredient-details-food-energy/modal-ingredient-details-food-energy.module.css
