+++ src/components/modal/modal-ingredient-details/modal-ingredient-details-image/modal-ingredient-details-image.module.css
