+++ src/components/modal/modal-order-creation-details/modal-order-creation-details.module.css
@@ -1,4 +1,4 @@
-@import '../../common/styles/variables.css';
+@import '../../../common/styles/variables.css';
 
 .dialog {
   display: flex;
