+++ src/components/modal/modal-order-details/modal-order-details.module.css
@@ -1,4 +1,4 @@
-@import './../../common/styles/variables.css';
+@import '../../../common/styles/variables.css';
 
 .dialog {
   display: flex;
